Move util interfaces to separate config package
