+++ cloudscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/cloudscrambles/KotlessCloudscramblesServer.kt
@@ -2,7 +2,7 @@ package org.worldcubeassociation.tnoodle.server.cloudscrambles
 
 import io.kotless.dsl.ktor.Kotless
 import io.ktor.application.Application
-import org.worldcubeassociation.tnoodle.server.LocalServerEnvironmentConfig
+import org.worldcubeassociation.tnoodle.server.config.LocalServerEnvironmentConfig
 import org.worldcubeassociation.tnoodle.server.cloudscrambles.CloudscramblesServer.wrapAndLaunchServer
 
 class KotlessCloudscramblesServer : Kotless() {
