+++ tnoodle-server/src/main/kotlin/org/worldcubeassociation/tnoodle/server/routing/VersionHandler.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.json.json
 
 import org.worldcubeassociation.tnoodle.server.RouteHandler
 import org.worldcubeassociation.tnoodle.server.cryptography.AsymmetricCipher
-import org.worldcubeassociation.tnoodle.server.util.ServerEnvironmentConfig
+import org.worldcubeassociation.tnoodle.server.ServerEnvironmentConfig
 
 class VersionHandler(val version: ServerEnvironmentConfig) : RouteHandler {
     override fun install(router: Route) {
