+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/WebscramblesServer.kt
@@ -11,11 +11,11 @@ import org.slf4j.LoggerFactory
 import org.worldcubeassociation.tnoodle.server.ApplicationHandler
 import org.worldcubeassociation.tnoodle.server.TNoodleServer
 import org.worldcubeassociation.tnoodle.server.routing.JsEnvHandler
-import org.worldcubeassociation.tnoodle.server.util.ServerEnvironmentConfig
+import org.worldcubeassociation.tnoodle.server.ServerEnvironmentConfig
 import org.worldcubeassociation.tnoodle.server.webscrambles.server.MainLauncher.NO_REEXEC_OPT
 import org.worldcubeassociation.tnoodle.server.webscrambles.routing.*
 import org.worldcubeassociation.tnoodle.server.webscrambles.routing.job.JobSchedulingHandler
-import org.worldcubeassociation.tnoodle.server.LocalServerEnvironmentConfig
+import org.worldcubeassociation.tnoodle.server.config.LocalServerEnvironmentConfig
 import org.worldcubeassociation.tnoodle.server.webscrambles.routing.frontend.ApplicationDataHandler
 import org.worldcubeassociation.tnoodle.server.webscrambles.routing.frontend.WcifDataHandler
 import org.worldcubeassociation.tnoodle.server.webscrambles.server.MainLauncher
