+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/server/OfflineJarUtils.kt
@@ -3,7 +3,7 @@ package org.worldcubeassociation.tnoodle.server.webscrambles.server
 import dorkbox.systemTray.SystemTray
 import dorkbox.systemTray.MenuItem
 import org.slf4j.LoggerFactory
-import org.worldcubeassociation.tnoodle.server.LocalServerEnvironmentConfig
+import org.worldcubeassociation.tnoodle.server.config.LocalServerEnvironmentConfig
 import java.awt.*
 import java.io.IOException
 import java.net.URI
